Tests/LibWeb/Text/expected/selection-drag-past-overflowing-text-and-image.txt

kept arrow selected below row: true
